// be/src/common/errors/error-codes.enum.ts
export enum ErrorCode {
  // Authentication & Authorization
  AUTH_INVALID_CREDENTIALS = 'AUTH_INVALID_CREDENTIALS',
  AUTH_TOKEN_EXPIRED = 'AUTH_TOKEN_EXPIRED',
  FORBIDDEN_ROLE_MISMATCH = 'FORBIDDEN_ROLE_MISMATCH',

  // Users
  USER_ALREADY_EXISTS = 'USER_ALREADY_EXISTS',
  USER_NOT_FOUND = 'USER_NOT_FOUND',

  // Clients & Reviewers
  CLIENT_ALREADY_EXISTS = 'CLIENT_ALREADY_EXISTS',
  CLIENT_NOT_FOUND = 'CLIENT_NOT_FOUND',
  REVIEWER_ROLE_REQUIRED = 'REVIEWER_ROLE_REQUIRED',
  REVIEWER_ALREADY_ASSIGNED = 'REVIEWER_ALREADY_ASSIGNED',
  REVIEWER_NOT_ASSIGNED_TO_CLIENT = 'REVIEWER_NOT_ASSIGNED_TO_CLIENT',

  // Posts Workflow & Business Rules
  POST_NOT_FOUND = 'POST_NOT_FOUND',
  INVALID_STATUS_TRANSITION = 'INVALID_STATUS_TRANSITION',
  SELF_APPROVAL_FORBIDDEN = 'SELF_APPROVAL_FORBIDDEN',
  CREATOR_EDIT_FORBIDDEN = 'CREATOR_EDIT_FORBIDDEN',
  POST_STATUS_NOT_EDITABLE = 'POST_STATUS_NOT_EDITABLE',
  CHANGES_REQUEST_COMMENT_REQUIRED = 'CHANGES_REQUEST_COMMENT_REQUIRED',
  CAPTION_LIMIT_EXCEEDED = 'CAPTION_LIMIT_EXCEEDED',
  SCHEDULED_TIME_NOT_FUTURE = 'SCHEDULED_TIME_NOT_FUTURE',
  SCHEDULED_TIME_REQUIRED = 'SCHEDULED_TIME_REQUIRED',

  // Concurrency & Scheduling Conflicts (409)
  OPTIMISTIC_LOCK_CONFLICT = 'OPTIMISTIC_LOCK_CONFLICT',
  SCHEDULING_CONFLICT_2HR = 'SCHEDULING_CONFLICT_2HR',
}